langs: move result to a higher level and reuse
