+++ src/tests/langs/javascript/generate_test.ts
@@ -1,7 +1,8 @@
 import * as assert from "assert";
 import { generateProgram } from "../../../langs/javascript/generate.ts";
 import { parse } from "../../../langs/javascript/parser/parse.ts";
-import type { Program, Result } from "../../../langs/javascript/types.ts";
+import type { Program } from "../../../langs/javascript/types.ts";
+import type { Result } from "../../../langs/types.ts";
 
 function expectOk<T>(result: Result<T>): T {
     if (result.kind !== "Ok") {
